+++ Location/Controller/Adminhtml/City/MassDelete.php
@@ -58,19 +58,23 @@ class MassDelete extends Action
      */
     public function execute()
     {
-        $deletedItemsCount = 0;
-        $collection = $this->massActionFilter->getCollection($this->cityCollectionFactory->create());
-        foreach ($collection as $city) {
-            try {
-                /** @var CityInterface $city */
-                $this->cityRepository->deleteById($city->getCityId());
-                $deletedItemsCount++;
-            } catch (CouldNotDeleteException $e) {
-                $errorMessage = __('[ID: %1] ', $city->getCityId()) . $e->getMessage();
-                $this->messageManager->addErrorMessage($errorMessage);
+        if ($this->getRequest()->isPost()) {
+            $deletedItemsCount = 0;
+            $collection = $this->massActionFilter->getCollection($this->cityCollectionFactory->create());
+            foreach ($collection as $city) {
+                try {
+                    /** @var CityInterface $city */
+                    $this->cityRepository->deleteById($city->getCityId());
+                    $deletedItemsCount++;
+                } catch (CouldNotDeleteException $e) {
+                    $errorMessage = __('[ID: %1] ', $city->getCityId()) . $e->getMessage();
+                    $this->messageManager->addErrorMessage($errorMessage);
+                }
             }
+            $this->messageManager->addSuccessMessage(__('You deleted %1 city(s).', $deletedItemsCount));
+        } else {
+            $this->messageManager->addErrorMessage(__('Wrong request.'));
         }
-        $this->messageManager->addSuccessMessage(__('You deleted %1 city(s).', $deletedItemsCount));
-        return $this->resultRedirectFactory->create()->setPath('*/*/');
+        return $this->resultRedirectFactory->create()->setPath('*/*/index');
     }
 }
